requirements.txt
SpeechRecognition==3.6.3
PyPDF2==1.26.0
gTTS==2.0.1
urllib3>=1.24.1
pdf2image==1.6.0
Pillow==5.4.1
pyfiglet==0.7
pytesseract
PyInquirer==1.0.3

